Ajustando modulo de usuário

+++ src/main/java/org/example/repository/IEntityRepository.java
@@ -7,6 +7,6 @@ public interface IEntityRepository<T> {
     public void save(T entity);
     public void delete(T entity);
     public void update(T entity);
-    public T findById(Long id);
+    public T findById(int id);
     public List<T> findAll();
 }
